+++ vite-project/index.html
@@ -28,11 +28,49 @@
         <button type="submit" id="submit-btn">Add</button>
       </div>
       <div class="list">
-        <div class="list-title"></div>
-        <div class="list-wrap"></div>
+        <div class="list-date" id="list-date"></div>
+        <div class="list-wrap">
+          <ul>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>과제하기</p>
+            </li>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>과제하기</p>
+            </li>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>과제하기</p>
+            </li>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>과제하기</p>
+            </li>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>과제하기</p>
+            </li>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>늦잠 자기</p>
+            </li>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>마라샹궈 먹기</p>
+            </li>
+            <li>
+              <input type="checkbox" name="todo" />
+              <p>악보보고 피아노 치기 완벽할 때 까지 코난</p>
+            </li>
+          </ul>
+        </div>
         <div class="list-btn">
-          <button id="list-edit">Edit</button>
-          <button id="list-delete">Delete</button>
+          <!-- <button id="list-edit">Edit</button> -->
+          <button id="list-delete">DELETE</button>
+        </div>
+        <div class="complete">
+          <p>2/3</p>
         </div>
       </div>
       <div class="button">
