+++ vite-project/js/selectDay.js
@@ -1,0 +1,15 @@
+// export default function selectDay() {
+//   let selectYear = date.getFullYear();
+//   let selectMonth = date.getMonth()+1;
+//   let selectDate = document.querySelectorAll('.this');
+
+//   const listDateEl = document.getElementById("list-date");
+//   for(let i = 0; i < selectDate.length; i++){
+//     selectDate[i].addEventListener("click", () => {
+//       // console 확인
+//       // console.log(selectYear, selectMonth, Number(selectDate[i].textContent));
+//       listDateEl.textContent = `${selectYear}.${selectMonth}.${Number(selectDate[i].textContent)}`
+//       return selectYear, selectMonth, Number(selectDate[i].textContent);
+//     })
+//   } 
+// }
