+++ vite-project/scss/main.scss
@@ -1,5 +1,6 @@
 
 body {
+  height: auto;
   background-color: #F5F5F5;
   display: flex;
 }
@@ -8,7 +9,6 @@ main {
   display: flex;
   justify-content: center;
   width: 40%;
-  height: 100vh;
   background-color: #CE2424;
   .container {
     .title {
@@ -30,23 +30,23 @@ main {
         width: 100%;
         height : 50px;
         border-radius: 0;
+        padding-top: 7px;
         border: none;
         padding-left: 20px;
         background-color: #F5F5F5;
         color: #CE2424;
-        font-size: 18px;
+        font-size: 23px;
         font-family: 'Black Han Sans','Erica One', cursive;
         box-sizing: border-box;
       }
       #submit-btn {
         position: absolute;
         right: 8px;
-        width: 80px;
+        width: 100px;
         height: 40px;
         border: none;
         color: #F5F5F5;
         letter-spacing: 1px;
-        border-radius: 3px;
         font-size: 18px;
         font-family: 'Erica One', cursive;
         background-color: #CE2424;
@@ -54,31 +54,66 @@ main {
       }
     }
     .list {
-      width: 100%;
+      width: 530px;
       height: 500px;
       margin-top: 10px;
+      box-sizing: border-box;
       background-color: #F5F5F5;
       display: flex;
       justify-content: center;
       align-content: flex-start;
       position: relative;
       flex-wrap: wrap;
-      .list-title {
+      .list-date {
+        text-align: center;
         width: 90%;
         height: 60px;
         margin-top: 30px;
         margin-bottom: 20px;
-        background-color: aquamarine;
+        font-size: 55px;
+        font-family: 'Erica One', cursive;
+        color: #CE2424;
+        // background-color: aquamarine;
       }
       .list-wrap {
         width: 90%;
         height: 320px;
         padding: 15px;
         box-sizing: border-box;
-        background-color: aquamarine;
+        overflow-x: auto;
+        // background-color: aquamarine;
+        li {
+          padding-top: 13px;
+          display: flex;
+          align-items: center;
+          border-bottom: 2px solid #CE2424;
+          padding-bottom: 10px;
+          [type="checkbox"] {
+            // vertical-align: middle;
+            appearance: none;
+            border: max(2px, 0.1em) solid #CE2424;
+            border-radius: 50%;
+            width: 1.25em;
+            height: 1.25em;
+            transition: 0.3s ease-in-out;
+          }
+          [type="checkbox"]:checked {
+            // border: 0.6em solid #CE2424
+            background-color: #CE2424;
+          }
+          p {
+            padding-top: 4px;
+            padding-left: 10px;
+            box-sizing: border-box;
+            font-size: 20px;
+            font-family: 'Black Han Sans','Erica One', cursive;
+            color: #CE2424;
+          }
+        }
       }
       .list-btn {
         position: absolute;
+        left: 20px;
         bottom: 20px;
         button {
           border: none;
@@ -86,6 +121,21 @@ main {
           color: #CE2424;
           font-family: 'Erica One', cursive;
           font-size: 30px;
+          padding: 0 20px;
+          text-align: center;
+          cursor: pointer;
+        }
+      }
+      .complete {
+        position: absolute;
+        right: 10px;
+        bottom: 20px;
+        p {
+          border: none;
+          background-color: transparent;
+          color: #CE2424;
+          font-family: 'Erica One', cursive;
+          font-size: 40px;
           padding: 0 20px;
           text-align: center;
           cursor: pointer;
